+++ examples/new_template_library.py
@@ -8,7 +8,8 @@ from astropy.io import fits
 from matplotlib import pyplot
 from matplotlib.ticker import NullFormatter
 
-from mangadap.config.defaults import dap_source_dir
+from mangadap.config import defaults
+from mangadap.config import manga
 from mangadap.datacube import MaNGADataCube
 from mangadap.util.constants import DAPConstants
 from mangadap.util.fitsutil import DAPFitsUtil
@@ -31,7 +32,7 @@ from mangadap.proc.emissionlinemodel import EmissionLineModelBitMask
 #-----------------------------------------------------------------------------
 def get_redshift(plt, ifu, drpall_file=None):
     if drpall_file is None:
-        drpall_file = defaults.drpall_file()
+        drpall_file = manga.drpall_file()
     with fits.open(drpall_file) as hdu:
         indx = hdu[1].data['PLATEIFU'] == '{0}-{1}'.format(plt, ifu)
         return hdu[1].data['NSA_Z'][indx][0]
@@ -80,7 +81,7 @@ if __name__ == '__main__':
     # Create a library based on a subset of an existing one
 
     # Add a new library that only has 10 Gyr SSPs from MIUSCAT
-    file_search = os.path.join(dap_source_dir(), 'data', 'spectral_templates', 'miuscat',
+    file_search = os.path.join(defaults.dap_source_dir(), 'data', 'spectral_templates', 'miuscat',
                                '*T10.0000.fits')
 
     # Define the template library
@@ -121,7 +122,7 @@ if __name__ == '__main__':
 
     # Get the redshift
     drpver = 'v2_7_1'
-    directory_path = os.path.join(dap_source_dir(), 'data', 'remote')
+    directory_path = os.path.join(defaults.dap_source_dir(), 'data', 'remote')
     drpall_file = os.path.join(directory_path, 'drpall-{0}.fits'.format(drpver))
     z = numpy.array([get_redshift(plt, ifu, drpall_file)])
     print('Redshift: {0}'.format(z[0]))
